1.3.0

+++ taxontabletools/create_metadata_table.py
@@ -4,7 +4,15 @@ def create_metadata_table(TaXon_table_xlsx, path_to_outdirs):
     import pandas as pd
     from pandas import DataFrame
     import numpy as np
+    import sys, subprocess, os
     from pathlib import Path
+
+    def open_table(table):
+        if sys.platform == "win32":
+            os.startfile(table)
+        else:
+            opener = "open" if sys.platform == 'darwin' else 'xdg-open'
+            subprocess.call([opener, table])
 
     TaXon_table_xlsx = Path(TaXon_table_xlsx)
     Meta_data_table_xlsx = Path(str(path_to_outdirs) + "/" + "Meta_data_table" + "/" + TaXon_table_xlsx.stem + "_metadata.xlsx")
@@ -43,15 +51,31 @@ def create_metadata_table(TaXon_table_xlsx, path_to_outdirs):
 
     window_progress_bar.Close()
 
-    metadata_df = pd.DataFrame(samples_metadata_list)
-    metadata_df.columns  = ["Samples"] + ["col_" + str(column) for column in metadata_df.columns.tolist()[1:]]
-    metadata_df.to_excel(Meta_data_table_xlsx, index=False)
+    answer = "No"
+    if Meta_data_table_xlsx.exists():
+        answer = sg.PopupYesNo("Metadata tables already exists! Overwrite?")
+        if answer == "Yes":
+            metadata_df = pd.DataFrame(samples_metadata_list)
+            metadata_df.columns  = ["Samples"] + ["col_" + str(column) for column in metadata_df.columns.tolist()[1:]]
+            metadata_df.to_excel(Meta_data_table_xlsx, index=False)
 
-    closing_text = "Metadata table is found under:\n" + '/'.join(str(Meta_data_table_xlsx).split("/")[-4:])
-    sg.Popup(closing_text, title="Finished", keep_on_top=True)
+            answer = sg.PopupYesNo("Open metadata table?", title="Finished", keep_on_top=True)
+            if answer == "Yes":
+                open_table(Meta_data_table_xlsx)
 
-    from taxontabletools.create_log import ttt_log
-    ttt_log("meta data table", "analysis", TaXon_table_xlsx.name, Meta_data_table_xlsx.name, "nan", path_to_outdirs)
+            from taxontabletools.create_log import ttt_log
+            ttt_log("meta data table", "analysis", TaXon_table_xlsx.name, Meta_data_table_xlsx.name, "nan", path_to_outdirs)
+    else:
+        metadata_df = pd.DataFrame(samples_metadata_list)
+        metadata_df.columns  = ["Samples"] + ["col_" + str(column) for column in metadata_df.columns.tolist()[1:]]
+        metadata_df.to_excel(Meta_data_table_xlsx, index=False)
+
+        answer = sg.PopupYesNo("Open metadata table?", title="Finished", keep_on_top=True)
+        if answer == "Yes":
+            open_table(Meta_data_table_xlsx)
+
+        from taxontabletools.create_log import ttt_log
+        ttt_log("meta data table", "analysis", TaXon_table_xlsx.name, Meta_data_table_xlsx.name, "nan", path_to_outdirs)
 
 def modify_metadata_table(path_to_outdirs):
     import PySimpleGUI as sg
